Fix: RegLog Feature | -minus: Animation

+++ app/src/main/java/com/example/capstone/view/viewmodel/RegisterViewModel.kt
@@ -6,7 +6,7 @@ import com.example.capstone.data.AppRepository
 class RegisterViewModel(private val repository: AppRepository): ViewModel() {
     fun register(
         name: String, username: String, email: String, password: String,
-        berat: Int, tinggi: Int, gulaDarah: Float, kolestrol: Float, tekanan: Float
+        berat: Float, tinggi: Float, gulaDarah: Float, kolestrol: Float, tekanan: Float
     ) = repository.register(
         name, username, email, password, berat, tinggi,
         gulaDarah, kolestrol, tekanan
